Shift the deafult button assignment to the dialog
The original design allowed a user to mark multiple buttons as default.
In this situation there will be a RTE, which is not the best
devex/UX.

Shift assignment of the default button to the dialog. This way the API
are similar to the MessageBox API, and the user is able to assign
onle one button.

If the button doesn't belong to the collection of buttons - there will
be a RTE.