Simplify setting dialog icon from bitmaps
Typically images are used from Resources, and those are of `Bitmap` type.
The existing API only takes `Icon` that requires additional work from a
developer, which wouldn't be straight forward for a significant part of
the development community, e.g.:

    new TaskDialogIcon(Icon.FromHandle(Properties.Resources.Network.GetHicon()))

Provide an overload that takes a `Bitmpa` and handles the necessary plumbing.